requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.52
pymysql==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pydantic-settings
numpy
scikit-learn
passlib
bcrypt
pandas 